# doc/api.yaml
openapi: 3.0.0
info:
  title: WasaText
  description: |
    This API allows users to interact with each other 
    on the WasaText app.
    Users can create groups, private chats, send text or image messages, 
    reply, forward or react them.
  version: 1.0.0
servers:
  - url: http://localhost:3000/v1
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: UUID
      description: |
        Use the user identifier returned by `/login` in the Authorization header.
        Example: `Authorization: Bearer abcdef012345`
  responses:
    UnauthorizedError:
      description: Missing or invalid Authorization header
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Unauthorized"
...
security:
  - BearerAuth: []
paths:
...
  /conversations:
    post:
      tags:
        - Conversations
      summary: Create a new conversation
      description: |
        Create a new private conversation.  
        If a private conversation with the specified user already exists, the existing one is returned.
      operationId: createConversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: The recipient id to start the conversation with
              oneOf:
                - type: object
                  required:
                    - recipientId
                  properties:
                    recipientId:
                      type: integer
                      description: The ID of the recipient user.
              example:
                recipientId: 23
      responses:
        "201":
          description: Conversation created successfully
          content:
            application/json:
              schema:
                description: An object for the newly created conversation
                type: object
                properties:
                  conversationId:
                    type: integer
                    description: The conversation id
                  type:
                    type: string
                    description: The conversation type 'private' | 'group'
                  createdBy:
                    type: integer
                    description: The user id of the conversation creator
                  createdAt:
                    type: string
                    format: date-time
                    description: The date of conversation creation
              example:
                conversationId: 18
                type: private
                createdBy: 31
                createdAt: "2025-10-12T15:42:00Z"
        "200":
          description: Existing conversation returned
          content:
            application/json:
              schema:
                description: An object fot the existing conversation
                type: object
                properties:
                  conversationId:
                    type: integer
                    description: The conversation id
                  type:
                    type: string
                    description: The conversation type 'private' | 'group'
                  createdBy:
                    type: integer
                    description: The user id of the conversation creator
                  createdAt:
                    type: string
                    format: date-time
                    description: The date of conversation creation
              example:
                conversationId: 7
                type: private
                createdBy: 31
                createdAt: "2025-09-25T09:20:00Z"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                description: A descriptive message of failure
                type: object
                properties:
                  message:
                    type: string
                    example: "Missing recipient id"
                    description: The message of failure
        "404":
          description: Recipient not found
          content:
            application/json:
              schema:
                description: A descriptive message of failure
                type: object
                properties:
                  message:
                    type: string
                    example: "User not found"
                    description: The message of failure
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    get:
      tags:
        - Conversations
      summary: Get a user's conversations
      description:
        Get all user's conversations. Each conversation entry will contain the conversation fields and
        if present, a lastMessage entry acting as a preview for the conversation.
      operationId: getMyConversations
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                description: An object containing the array of conversations for the user
                type: object
                properties:
                  conversations:
                    type: array
                    items:
                      type: object
                      properties:
                        conversationId:
                          type: integer
                          description: The conversation id
                        type:
                          type: string
                          enum: [private, group]
                          description: The conversation type 'private' | 'group'
                        name:
                          type: string
                          description: The conversation name (if the conversation is a group)
                        photoPath:
                          type: string
                          description: The conversation photo path (if the conversation is a group)
                        createdBy:
                          type: integer
                          description: The conversation creator's user id
                        createdAt:
                          type: string
                          format: date-time
                          description: The conversation datetime of creation
                        lastMessage:
                          type: object
                          description: The last message sent in the conversation
                          properties:
                            id:
                              type: integer
                              description: The message id
                            type:
                              type: string
                              enum: [text, image]
                              description: The message type 'text' | 'image'
                            isForwarded:
                              type: boolean
                              description: Whether the message was forwarded
                            authorId:
                              type: integer
                              description: The user id that created the message
                            content:
                              type: string
                              description: The message text content (for type text) or photo path (for type image)
                            createdAt:
                              type: string
                              format: date-time
                              description: The message datetime of creation
              example:
                conversations:
                  - conversationId: 7
                    type: group
                    name: Partite Calcio
                    photoPath: "groups/7/photo.png"
                    createdBy: 31
                    createdAt: "2025-09-29T14:48:00Z"
                    lastMessage:
                      id: 239
                      type: text
                      isForwarded: false
                      authorId: 20
                      content: "Va bene ci sono per stasera, ci vediamo!"
                      createdAt: "2025-09-28T16:30:00Z"
                  - conversationId: 12
                    type: private
                    name: "Edoardo"
                    photoPath: null
                    createdBy: 31
                    createdAt: "2025-09-25T09:20:00Z"
                    lastMessage:
                      id: 312
                      type: text
                      isForwarded: false
                      authorId: 45
                      content: "Ci sentiamo più tardi!"
                      createdAt: "2025-09-28T18:10:00Z"

        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /conversations/{id}:
    get:
      tags:
        - Conversations
      summary: Get a conversation by id
      description:
        Get a conversation with all the messages belonging to it. The conversation can either be a
        private chat or a group chat. Each message entry will contain the message fields and
        the message status fields.
      operationId: getConversation
      parameters:
        - name: id
          in: path
          required: true
          description: The conversation id
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                description: An object containing an array of messages for the conversation
                type: object
                properties:
                  messages:
                    description: Array of messages for the conversation
                    type: array
                    items:
                      type: object
                      properties:
                        messageId:
                          type: integer
                          description: The message id
                        type:
                          type: string
                          enum: [text, image]
                          description: The message type 'text' | 'image'
                        isForwarded:
                          type: boolean
                          description: Wheather the message was forwarded
                        content:
                          type: string
                          description: The message text content (for type text) or photo path (for type image)
                        authorId:
                          type: integer
                          description: The message author id
                        createdAt:
                          type: string
                          format: date-time
                          description: The message datetime of creation
                        messageStatus:
                          type: object
                          description: The message status 'received' | 'read' with the id of the members who read it
                          properties:
                            status:
                              type: string
                              enum: [received, read]
                              description: The status of the message, 'read' if all the members read the message, 'received' otherwise
                            members:
                              description: The member ids that read the message
                              type: array
                              items:
                                type: integer
                        comments:
                          type: array
                          description: List of emoji reactions to this message
                          items:
                            type: object
                            properties:
                              authorId:
                                type: integer
                                description: The user id of the user that reacted
                              emoji:
                                type: string
                                description: The emoji reaction (Unicode)
                              createdAt:
                                type: string
                                format: date-time
                                description: The comment datetime of creation

              example:
                messages:
                  - messageId: 65
                    type: image
                    isForwarded: true
                    content: "groups/7/images/photo123.jpeg"
                    authorId: 3
                    createdAt: "2025-09-29T14:48:00Z"
                    messageStatus:
                      status: read
                      members: [1, 2, 3]
                    comments:
                      - authorId: 1
                        emoji: "🔥"
                        createdAt: "2025-09-29T14:48:32Z"
                      - authorId: 2
                        emoji: "😂"
                        createdAt: "2025-09-29T14:49:00Z"
                  - messageId: 66
                    type: text
                    isForwarded: false
                    content: "AHHAAHH che ridere!"
                    authorId: 1
                    created_at: "2025-09-29T15:00:00Z"
                    messageStatus:
                      status: received
                      members: [1, 3]
                    comments: []
                  - messageId: 67
                    type: text
                    isForwarded: false
                    content: "No vabbe troppo divertente"
                    authorId: 2
                    createdAt: "2025-09-29T15:05:00Z"
                    messageStatus:
                      status: received
                      members: [3]
                    comments: []
        "401":
          $ref: "#/components/responses/UnauthorizedError"

        "403":
          description: The user doesn't belong to the conversation
          content:
            application/json:
              schema:
                description: A descriptive messafe of failure
                type: "object"
                properties:
                  message:
                    type: "string"
                    description: The user doesn't belong to this conversation.
              example:
                message: User doesn't belong to the conversation
...